requirements.txt
pyspark==3.1.1
pandas==1.2.4
dbx==0.3.0
nltk==3.6.2
spacy==3.2.4
numpy==1.20.2
pyarrow==3.0.0
iab-tcf==0.1.7
boto3==1.15.0
unidecode==1.2.0
pydantic==1.10.11
moto==1.3.16
slack_sdk==3.11.0
pymysql==1.0.2
scikit-learn==1.0.2
pre-commit==2.15.0
aenum==3.1.5
matplotlib==3.4.3
mlflow==1.15.0
snowflake-connector-python==2.7.4
torch==1.10.1
transformers==4.15.0
sentencepiece==0.1.96
faiss-cpu==1.7.2
protobuf==3.20.0
typing-extensions==4.7.1